requirements.txt
asgiref==3.7.2
Django==4.2.7
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==23.4
django-imagekit==5.0.0
djangorestframework==3.14.0
pilkit==3.0
Pillow==10.1.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
